work please

+++ web/templates/category_form.php
@@ -130,6 +130,7 @@
 								<div class="col-sm-12 form-group">
 									<button type="submit" class="button2" >Send &rarr;</button>
 									<?php
+									if (!$is_create)
 										echo "<a class='button2' id= 'mylink' href='inccats.php?delete_data=".$id."'>DELETE</a>";
 									?>
 								</div>
